requirements.txt
PyPDF2>=3.0.0
ollama-python>=0.1.0
google-generativeai>=0.4.0
streamlit>=1.30.0
